requirements.txt
Requirements
Python Version:
Python 3.12.4
Libraries/Modules:
iPython
requests
PyGithub








